First step of playing music in the UI.
In order to seperate the UI logic from the player logic, I introduce the
UiController (which might get a more suitable name later on). This
UiController serves here as a factory for PlayWidget, enabling it
to gives instances of PlayWidgetEntryController to newly created
PlayWidgetEntry elements.

PlayWidgetEntryController takes care of calling Players to play
the Track.